+++ CHANGES.md
@@ -11,6 +11,8 @@
  - Fix broken `coq/saveVo` and `coq/getDocument` requests due to a
    parsing problem with extra fields in their requests (@ejgallego,
    #547, reported by @Zimmi48)
+ - `fcc` now understands the `--coqlib`, `--coqcorelib`,
+   `--ocamlpath`, `-Q` and `-R` arguments (@ejgallego, #555)
 
 # coq-lsp 0.1.7: Just-in-time
 -----------------------------